+++ src/Colony.ts
@@ -32,6 +32,7 @@ import {$} from './caching/GlobalCache';
 import {DirectiveHarvest} from './directives/resource/harvest';
 import {DirectiveExtract} from './directives/resource/extract';
 import {Cartographer, ROOMTYPE_CONTROLLER} from './utilities/Cartographer';
+import {Visualizer} from './visuals/Visualizer';
 
 export enum ColonyStage {
 	Larva = 0,		// No storage and no incubator
@@ -540,7 +541,75 @@ export class Colony {
 		}
 	}
 
+	private defconReport(): void {
+		// let safeOutposts = _.filter(this.outposts, room => !!room && room.dangerousHostiles.length == 0);
+		// let stringReport: string[] = [
+		// 	`DEFCON: ${this.defcon}  Safe outposts: ${safeOutposts.length}/${this.outposts.length}`,
+		// 	`Creep usage for ${colony.name}:`];
+	}
+
+	private drawStorageReport(y = 11.5): number {
+		if (!this.commandCenter) return y;
+		let height = this.commandCenter.storage && this.commandCenter.terminal ? 2 : 1;
+		let boxCoords = Visualizer.section(`${this.name} Command Center`, {x: 1, y: y, roomName: this.room.name},
+										   9.5, height + .1);
+		let boxX = boxCoords.x;
+		y = boxCoords.y + 0.25;
+		if (this.commandCenter.storage) {
+			Visualizer.text('Storage', {x: boxX, y: y, roomName: this.room.name});
+			Visualizer.barGraph(_.sum(this.commandCenter.storage.store) / this.commandCenter.storage.storeCapacity,
+				{x: boxX + 4, y: y, roomName: this.room.name}, 5);
+			y += 1;
+		}
+		if (this.commandCenter.terminal) {
+			Visualizer.text('Terminal', {x: boxX, y: y, roomName: this.room.name});
+			Visualizer.barGraph(_.sum(this.commandCenter.terminal.store) / this.commandCenter.terminal.storeCapacity,
+				{x: boxX + 4, y: y, roomName: this.room.name}, 5);
+			y += 1;
+		}
+		return y + .25;
+	}
+
+	private drawCreepReport(y = 11.5): number {
+		const roledata = Overmind.overseer.getCreepReport(this);
+		const tablePos = new RoomPosition(1, y, this.room.name);
+		return Visualizer.infoBox(`${this.name} Creeps`, roledata, tablePos, 7);
+	}
+
+	private drawSpawnReport(y = 11.5): number {
+		if (!this.hatchery) return y;
+		let spawning: string[] = [];
+		let spawnPercents: number[] = [];
+		_.forEach(this.hatchery.spawns, function (spawn) {
+			if (spawn.spawning) {
+				spawning.push(spawn.spawning.name.split('_')[0]);
+				spawnPercents.push(1 - spawn.spawning.remainingTime / spawn.spawning.needTime);
+			}
+		});
+		let boxCoords = Visualizer.section(`${this.name} Hatchery`, {x: 1, y: y, roomName: this.room.name},
+										   9.5, 1 + spawning.length + .1);
+		let boxX = boxCoords.x;
+		y = boxCoords.y + 0.25;
+		let uptime = this.hatchery.memory.stats.uptime;
+		Visualizer.text('Uptime', {x: boxX, y: y, roomName: this.room.name});
+		Visualizer.barGraph(uptime, {x: boxX + 4, y: y, roomName: this.room.name}, 5);
+		y += 1;
+		for (let i in spawning) {
+			Visualizer.text(spawning[i], {x: boxX, y: y, roomName: this.room.name});
+			Visualizer.barGraph(spawnPercents[i], {x: boxX + 4, y: y, roomName: this.room.name}, 5);
+			y += 1;
+		}
+		return y + .25;
+	}
+
 	visuals(): void {
+		let y = 11.5;
+		y = this.drawCreepReport(y);
+		y = this.drawSpawnReport(y);
+		y = this.drawStorageReport(y);
+		// Draw creep report
+
+
 		// let report: string[] = [];
 		// report.push(`Hatchery:`)
 		// for (let spawn of this.spawns) {